==> notebooks/4 - Categorization.ipynb <==
import pandas as pd

places_df = pd.read_csv('../data/places_GP_temp.csv')

places_df.title.fillna('', inplace = True)
places_df.category.fillna('', inplace = True)
places_df['new_category'] = ''

# Категории

Можно выделить доп категории: магазин, здоровье, машины, еда, школа, услуги

places_df['types'].apply(lambda s: list(ast.literal_eval(s))).explode().str.split('_').explode().value_counts().reset_index().to_csv('../data/types.csv', index = False)

places_df.loc[places_df.category != '','category'].str.lower().explode().str.split(' ').explode().value_counts().reset_index().to_csv('../data/category.csv', index = False)

places_df['types'].apply(lambda s: list(ast.literal_eval(s))).explode().value_counts().reset_index().to_csv('../data/types_GP.csv')

places_df.loc[places_df.category != '','category'].str.lower().value_counts().reset_index().to_csv('../data/category_GP.csv', index = False)

Услуга

places_df.loc[(places_df.types.str.contains('car_repair|beauty_salon|hair_care|real_estate_agency|travel_agency|car_dealer|spa|florist|car_rental|car_wash|plumber|electrician|moving_company|laundry|lawyer|funeral_home|post_office|locksmith|painter|insurance_agency|gas_station|bank')), 'new_category'] \
    = "Услуга"

Магазины

places_df.loc[((places_df.title.str.lower().str.contains('магазин'))|\
               (places_df.category.str.lower().str.contains('магазин'))|\
               (places_df.types.str.contains('store|shop')))&\
              ~((places_df.types.str.contains('bakery'))|\
                (places_df.types.str.contains('restaurant'))|\
                (places_df.types.str.contains('cafe'))|\
                (places_df.types.str.contains('bar'))), 'new_category'] = "Магазин"

Образовательные учереждения

places_df.loc[(places_df.types.str.contains('school|university|secondary_school|primary_school'))&\
              (places_df.category !="Магазин фермерских товаров"), 'new_category'] = "Образование"

Здоровье

places_df.loc[places_df.types.str.contains('health'), "new_category"] = "Здоровье"

Спорт

Можем говорить о том, что все спортзалы (gym) относятся к категории здоровья

places_df.loc[places_df.types.str.contains('gym'), 'types'].apply(lambda s: list(ast.literal_eval(s))).explode().str.split('_').explode().value_counts()[:20]

places_df.loc[places_df.types.str.contains('stadium')]

places_df.loc[(places_df.types.str.contains('gym'))|(places_df.types.str.contains('stadium')), 'new_category'] = "Спорт"

Общепит

places_df.loc[((places_df.types.str.contains('bakery'))|\
               (places_df.types.str.contains('restaurant'))|\
               (places_df.types.str.contains('cafe'))|\
               (places_df.types.str.contains('bar'))), 'new_category'] = "Общепит"

Остановка

places_df.loc[places_df.category.str.lower().str.contains('остановка'), 'category'].value_counts()

places_df.loc[places_df.category.str.lower().str.contains('остановка'), 'new_category'] = "Остановка ОТ"

ЖК

places_df.loc[places_df.title.str.lower().str.contains('жк ')].types.apply(lambda s: list(ast.literal_eval(s))).explode().value_counts()

places_df.loc[(places_df.title.str.lower().str.contains('жк '))&\
              ~(places_df.category.str.lower().str.contains('автобусная остановка')) |\
             (places_df.category.str.lower().str.contains('жилой комплекс')), 'category'].value_counts()

places_df.loc[(places_df.title.str.lower().str.contains('жк '))|\
             (places_df.category.str.lower().str.contains('жилой комплекс')), 'new_category'] = "Жилой комплекс"

Соотношение русский и англ категорий

gt = pd.read_csv('../data/groups_types.csv')
types = pd.read_csv('../data/types_GP.csv')

gt[84:]

', '.join(types.loc[gt[84:]['идентификатор'], 'index'].to_list())

places_df.loc[(~places_df.types.str.contains('car'))&\
              (places_df.types.str.contains('bus_station|light_rail_station|transit_station'))]
#               (places_df.types.str.contains('|'.join(types.loc[gt[82:84]['идентификатор'], 'index'].to_list())))]

Локальности

places_df.loc[(places_df.types.str.contains('locality|political|neighborhood|sublocality|sublocality_level_2|sublocality_level_1|airport|cemetery')), 'new_category'] \
    = "Локальность" 

открытые общественные пространства = 
Парк+Сквер+Бульвар(40), Сад (6)
площадь(10)
набережная (1)
Озеро (32)
пляж (4):
'tourist_attraction, park, natural_feature, town_square, amusement_park, zoo'

Помещения культурные(47)
театр+музей+галерея+церковь, собор+библиотека:

развлекательные
кинотеатр + боулинг
парк развлечений, 
зоопарк,

'bowling_alley, art_gallery, movie_theater, place_of_worship, library'

Зеленые зоны:

парк, сад, ботанический сад, экологический парк, сквер, бульвар

places_df.loc[(places_df.types.str.contains("'park'")) & ~(places_df.types.str.contains('car'))].category.value_counts()

places_df.loc[((places_df.category.str.contains('Парк|Сад|Ботанический сад|Экологический парк'))&\
             ~(places_df.category.str.contains('аттракционов|развлечений')) | \
              (places_df.title.str.lower().str.contains('сквер'))) & (places_df.new_category=='')]

places_df.loc[[253, 15198, 15283, 15291, 15311], 'new_category'] = "Сад"
places_df.loc[[15307,15305,10855,15250, 15304,15237,15209,15248,15303, 15302, 15301,15298,15297, 15296,15295,15294, 15290,15289,15287], 'new_category'] = "Сквер"
places_df.loc[[15300, 15308,15285], 'new_category'] = "Бульвар"
places_df.loc[[15309, 15299, 15281,15269,15203,15200,14197,6359,4419,3031,2905,1791,1610], 'new_category'] = "Парк"

places_df.loc[7765, 'new_category'] = "Образование"

places_df.loc[[2866,634, 645], 'new_category'] = "Зоопарк"

places_df.loc[(places_df.types.str.lower().str.contains('bowling_alley')), 'new_category'] = 'Боулинг'

places_df.loc[[687, 15462], 'new_category'] = "Кинотеатр"

places_df.loc[(places_df.types.str.lower().str.contains('art_gallery')), 'new_category'] = "Галерея"

places_df.loc[(places_df.category.str.lower().str.contains('музей'))|\
             (places_df.types.str.lower().str.contains('museum'))|\
             (places_df.title.str.lower().str.contains('музей')), 'new_category'] = "Музей"

places_df.loc[(places_df.category.str.lower() == 'театр'), 'new_category'] = "Театр"

places_df.loc[(places_df.types.str.lower().str.contains('library')), 'new_category'] = "Библиотека"

places_df.loc[(places_df.types.str.lower().str.contains('natural')), 'new_category'] = "Озеро"
places_df.loc[[15350,15273], 'new_category'] = "Озеро"

places_df.loc[[15368, 15340, 15333, 2934, 2764],'new_category'] = "Пляж"

places_df.loc[(places_df.types.str.lower().str.contains('place_of_worship|chur')), 'new_category'] = 'Религия'

places_df.loc[((places_df.category.str.lower().str.contains('церковь|храм|собор|религ'))|\
              (places_df.title.str.lower().str.contains('церковь|храм|собор|религ')))&
             (places_df.new_category==''), "new_category"] = "Религия"

places_df.loc[(places_df.category.str.lower().str.contains('аттрак'))|
             (places_df.category.str.lower().str.contains('развле'))|
             (places_df.title.str.lower().str.contains('аттрак')), 'new_category'] = 'Развлечение'
places_df.loc[9004, 'new_category'] = 'Развлечение'

places_df.loc[[674,2741, 3039, 14834,15306, 15310], 'new_category'] = "Аттракционы"

print(places_df.loc[15300, 'url'])

Достопримечательность=

памятник, достопримечательность, культурный объект, дома, фонтаны, деревья

places_df.loc[((places_df.category.str.lower().str.contains('памятник|мемори'))) & \
              ~(places_df.category.str.lower().str.contains('изготовитель')), 'new_category'] = "Памятник"

places_df.loc[(places_df.title.str.lower().str.contains('достопримечательность'))|\
              (places_df.category.str.lower().str.contains('достопримечательность')), 'new_category'] = 'Достопримечательность'

Набережная

places_df.loc[(places_df.title.str.lower().str.contains('набережная'))]

places_df.loc[[15382, 15225, 15445, 15451], 'new_category'] = "Набережная"

Площадь

places_df.loc[(places_df.category.str.lower().str.contains('площадь')), 'new_category'] = "Площадь" 

Мост

places_df.loc[(places_df.title.str.lower().str.contains("^мост|мост | мост"))|(places_df.category.str.lower() == "мост")]

places_df.loc[[15474, 15450, 15417,15398, 15373,  15345, 15338, 15335,15235,15219,10480,7563,7557,2814],'new_category'] = "Мост"

places_df.loc[places_df.new_category=='', 'types'].apply(lambda s: list(ast.literal_eval(s))).explode().value_counts()

places_df.loc[(places_df.new_category=='')& (places_df.types.str.contains('finance'))].category.value_counts()

places_df.loc[(places_df.new_category=='')& (places_df.types.str.contains('lodging')), 'new_category'] = 'Жилье'

places_df.loc[(places_df.types.str.contains('cloth')), 'new_category'] = 'Магазин одежды'

places_df.loc[(places_df.new_category=='')& (places_df.types.str.contains('tourist_attraction')), 'new_category'] =  'Культура'

places_df.loc[(places_df.new_category=='')& (places_df.types.str.contains('bus_station')), 'new_category'] = "Остановка ОТ"

places_df.loc[places_df.category.str.contains('Доставка готовой еды|Фреш-бар|Доставка пиццы|Ресторан или кафе|Кейтеринг|Пивоварня|Винодельня|Кафе-кондитерская'), 'new_category'] = 'Общепит'

places_df.new_category.drop_duplicates()

places_df.loc[10610, 'new_category'] = "Магазин"

places_df['new_category_2'] = places_df.new_category

places_df.loc[places_df.new_category.isin(['Магазин', 'Магазин одежды']),'new_category_2'] = "Магазин"
places_df.loc[places_df.new_category.isin(['Жилье', 'Услуга']),'new_category_2'] = "Услуга"
places_df.loc[places_df.new_category.isin(['Религия', 'Музей', "Библиотека", "Театр", "Галерея"]),'new_category_2'] = "Культура"
places_df.loc[places_df.new_category.isin(['Жилье', 'Услуга']),'new_category_2'] = "Услуга"
places_df.loc[places_df.new_category.isin(['Сад', 'Парк', "Бульвар", "Сквер"]),'new_category_2'] = "Зеленые зоны"
places_df.loc[places_df.new_category.isin(['Достопримечательность',"Памятник"]),'new_category_2'] = "Достопримечательность"
places_df.loc[places_df.new_category.isin(['Кинотеатр', 'Развлечение', "Аттракционы", "Зоопарк", "Боулинг"]),'new_category_2'] = "Развлечения"

places_df.new_category_2.value_counts()

places_df.drop([0, 13819],inplace = True)

# Сумма нормированных параметров
кво фото, кво отзывов, рейтинг

places_df = pd.read_csv('../data/places_GP.csv')

places_df.describe()
# .plot(kind='box', ylim = [0, 200])

places_df_scale = places_df.copy()

places_df_scale['count_review'] = places_df_scale['count_review'].apply(lambda x: x if x <=23 else 23)
places_df_scale['count_photo'] = places_df_scale['count_photo'].apply(lambda x: x if x <=27 else 27)

places_df['rating_by_category'] = 0
from sklearn.preprocessing import MinMaxScaler, RobustScaler
scaler = MinMaxScaler()
places_df['normalized_rating'] = scaler.fit_transform(places_df[['count_review', 'rating', 'count_photo']].fillna(0)).sum(axis = 1)
for i in places_df.category.value_counts().keys():
    places_df.loc[places_df.category==i, 'rating_by_category'] = scaler.fit_transform(places_df.loc[places_df.category==i][['count_review', 'rating', 'count_photo']].fillna(0)).sum(axis = 1)

places_df.sort_values('rating_by_category').loc[places_df.rating_by_category==3]

places_df.to_csv('../data/places_GP.csv', index = False)
