==> notebooks/6 - Collecting places reviews.ipynb <==
import pandas as pd
import requests
from multiprocessing.dummy import Pool
from multiprocessing import cpu_count, freeze_support
import shapely
from shapely import geometry
import re
import time
from selenium import webdriver
from selenium.webdriver.common.action_chains import ActionChains

#логер
def log_progress(sequence, every=None, size=None):
    from ipywidgets import IntProgress, HTML, VBox
    from IPython.display import display
    
    is_iterator = False
    if size is None:
        try:
            size = len(sequence)
        except TypeError:
            is_iterator = True
    if size is not None:
        if every is None:
            if size <= 200:
                every = 1
            else:
                every = size / 200     # every 0.5%
    else:
        assert every is not None, 'sequence is iterator, set every'

    if is_iterator:
        progress = IntProgress(min=0, max=1, value=1)
        progress.bar_style = 'info'
    else:
        progress = IntProgress(min=0, max=size, value=0)
    label = HTML()
    box = VBox(children=[label, progress])
    display(box)
    
    index = 0
    try:
        for index, record in enumerate(sequence, 1):
            
            if index == 1 or index % every == 0:
                if is_iterator:
                    label.value = '{index} / ?'.format(index=index)
                else:
                    progress.value = index
                    label.value = u'{index} / {size}'.format(
                        index=index,
                        size=size
                    )
            yield record
    except:
        progress.bar_style = 'danger'
        raise
    else:
        progress.bar_style = 'success'
        progress.value = index
        label.value = str(index or '?')

#comments
def get_comment_element(url, driver, count_review):
    driver.get(url)
    time.sleep(5)
    try:
        item = driver.find_element_by_css_selector('button[jsaction="pane.reviewChart.moreReviews"]')
        driver.execute_script("arguments[0].click();", item)
        return get_all_comments_place(url, driver, count_review)
    except:
        return 
    
def scroll(driver, path, count):
    items = []
    flag = True
    try:
        item = driver.find_element_by_xpath("//div[@class = 'siAUzd-neVct section-scrollbox cYB2Ge-oHo7ed cYB2Ge-ti6hGc']")
        last_height = driver.execute_script("return arguments[0].scrollHeight", item)
    except:
        flag = False


    while flag:
#         try:
#             # Scroll down to bottom
#             driver.execute_script("arguments[0].scrollTo(0, arguments[0].scrollHeight)", item)
#             new_height = driver.execute_script("return arguments[0].scrollHeight", item)
#         except:
#             new_height = last_height
#             pass
        new_height = last_height
        # Wait to load page
        start_time = time.time()
        while new_height == last_height:
            try:
                driver.execute_script("arguments[0].scrollTo(0, arguments[0].scrollHeight)", item)
                item = driver.find_element_by_xpath("//div[@class = 'siAUzd-neVct section-scrollbox cYB2Ge-oHo7ed cYB2Ge-ti6hGc']")
                # Calculate new scroll height and compare with last scroll height
                new_height = driver.execute_script("return arguments[0].scrollHeight", item)          
                items = driver.find_elements_by_xpath(path)
                if (len(items) >= count) or (time.time()-start_time)>35:
                    flag = False
                    break
            except Exception as e:
                print('3 '+e)
                flag = False
                break
        last_height = new_height

    items = driver.find_elements_by_xpath(path) 
    return items
    
def get_all_comments_place(place_url, driver, count_review):
    
    time.sleep(6)
    
    reviews_obj_list = []

    reviews = scroll(driver, '//div[@class="ODSEW-ShBeI NIyLF-haAclf gm2-body-2"]', count_review)
    
    for review in log_progress(reviews, 1):

#         нажать еще
        try:
            btn = review.find_element_by_xpath("//jsl /button")
            driver.execute_script("arguments[0].click();", btn)
        except:
            pass
        
        id_review = ''
        try:    
            id_review = review.get_attribute('data-review-id')
        except:
            pass
        
        photos = []
        try:
            p = review.find_elements_by_class_name('ODSEW-ShBeI-xJzy8c')
            if p[-1].get_attribute('aria-label')[0]=='+':
                driver.execute_script("arguments[0].click();", p[-1])
                time.sleep(1)
                p = review.find_elements_by_class_name('ODSEW-ShBeI-xJzy8c')
            photos = [re.search(r'\"https://\w.*\"', i.get_attribute('style')).group(0).replace('"', '') for i in p]
        except:
            pass

        rating = ""
        try:
            rating = int(review.find_element_by_class_name("ODSEW-ShBeI-H1e3jb").get_attribute('aria-label')[1])
        except:
            pass

        pub_date_text = ""
        try:    
            pub_date_text = review.find_element_by_class_name("ODSEW-ShBeI-RgZmSc-date").text
        except:
            pass

        
        reviewer_level = ''
        reviewer_counts = 0
        reviewer_page = ""
        try:
            reviewer = review.find_element_by_class_name("ODSEW-ShBeI-tXxcle-SfQLQb-menu")
            reviewer_page = reviewer.find_element_by_tag_name('a').get_attribute('href')
            level_and_counts = reviewer.text.replace('\n', '·').split('·')[1:]
            reviewer_level = level_and_counts[0].strip()
            reviewer_counts = level_and_counts[1].strip().split()[0]
        except:
            pass

        text = ""
        try:
            text = review.find_element_by_class_name("ODSEW-ShBeI-text").text
        except:
            pass   
        
        count_supports = 0
        try:
            count_supports = int(review.find_element_by_class_name("ODSEW-ShBeI-hWJfub-yHKmmc-NnAfwf").text)
        except:
            pass 

        obj = {
                'url': place_url,
                'id_review': id_review,
                'reviewer_page': reviewer_page,
                'reviewer_level': reviewer_level,
                'reviewer_counts': reviewer_counts,
                'rating': rating,
                'likes':count_supports,
                'date': pub_date_text,
                'text':text,
                'photos': photos
        }

        reviews_obj_list.append(obj)
    
    return reviews_obj_list

places_df = pd.read_csv('../data/places_GP.csv')
comments_df = pd.read_csv('../data/comments.csv')

temp_df = comments_df.groupby('url').count()['text'].reset_index()
comments_df = comments_df.loc[comments_df.url.isin(temp_df.loc[temp_df.text>8, 'url'])]
places_df.loc[(~places_df.count_review.isna()) & ~places_df.url.isin(comments_df.url)]

d = {0:[],1:[],2:[],3:[],4:[],5:[],6:[]}
for i, r in places_df.loc[(~places_df.count_review.isna()) & ~places_df.url.isin(comments_df.url)].sort_values('count_review', ascending = False, ignore_index=True).iterrows():
    d[i%7].append((r['url'], r['count_review']))

def get_reviews(urls):
    driver = webdriver.Chrome('chromedriver.exe')
    for url, count in log_progress(urls, 1):
        time.sleep(3)
        try:
            comments.extend(get_comment_element(url, driver, count))
        except:
            pass

comments = []

if __name__== "__main__":
    freeze_support()
    pool = Pool()
    ITERATION_COUNT = cpu_count()-1
    for i in range(0, ITERATION_COUNT):
        pool.apply_async(get_reviews, [d[i]])

c_df = pd.DataFrame(comments)

pd.concat([comments_df, c_df])

pd.concat([comments_df, c_df]).to_csv('../data/comments.csv', index = False)

comments_df
