==> notebooks/7 - Text analysis.ipynb <==
!pip install nltk

import pandas as pd #работа с датафреймами- таблицами
import pymorphy2 as pm #части речи, формы
import re #регулярные выражения
from multiprocessing.dummy import Pool
from multiprocessing import cpu_count, freeze_support
import time
from IPython.display import display 
import nltk

#логер
def log_progress(sequence, every=None, size=None):
    from ipywidgets import IntProgress, HTML, VBox
    from IPython.display import display
    
    is_iterator = False
    if size is None:
        try:
            size = len(sequence)
        except TypeError:
            is_iterator = True
    if size is not None:
        if every is None:
            if size <= 200:
                every = 1
            else:
                every = size / 200     # every 0.5%
    else:
        assert every is not None, 'sequence is iterator, set every'

    if is_iterator:
        progress = IntProgress(min=0, max=1, value=1)
        progress.bar_style = 'info'
    else:
        progress = IntProgress(min=0, max=size, value=0)
    label = HTML()
    box = VBox(children=[label, progress])
    display(box)
    
    index = 0
    try:
        for index, record in enumerate(sequence, 1):
            
            if index == 1 or index % every == 0:
                if is_iterator:
                    label.value = '{index} / ?'.format(index=index)
                else:
                    progress.value = index
                    label.value = u'{index} / {size}'.format(
                        index=index,
                        size=size
                    )
            yield record
    except:
        progress.bar_style = 'danger'
        raise
    else:
        progress.bar_style = 'success'
        progress.value = index
        label.value = str(index or '?')

places_df = pd.read_csv('../data/places_GP.csv')
df = pd.read_csv('../data/comments_filled.csv')
words = pd.read_csv('../data/words.csv')
bigramms = pd.read_csv('../data/bigramms.csv')

df = pd.read_csv('../data/comments.csv')
df = df.loc[~df.text.isna()]
display(df.sample())
df.shape

df.loc[df.text.str.contains('Google'), 'text'] = df.loc[df.text.str.contains('Google'), 'text'].astype(str).str.split('\n\n\(Оригинал\)\n', expand = False).str[0].str.split('\(Переведено Google\)\s').str[-1]

words = df['text'].str.extractall(r"(\S+)").reset_index()

# переименовываем колонки
words.rename(columns = {'level_0':'index_review', 0:'word'}, inplace = True)

# все буквы в словах делаем строчными
words.word = words.word.str.lower()

# создание новых столбцов, в нашем случае - индекса отзывов, 
# чтобы по этому индексу мы могли идентифицировать пренадлежность слов ко всем полям df
df['index_review']= df.index

# удаляем лишние символы, оставляя только слова
chars_to_remove = ['\r', '\n', '.', ')', '!', '(', ',', '-', ':', '?', '/', '"', ';']
regular_expression = '[' + re.escape (''. join (chars_to_remove)) + ']'
words.word = words.word.str.replace(regular_expression, ' ')

words = pd.DataFrame(words.word.str.strip().str.split().explode()).merge(words[['index_review', 'match']], how = 'left',left_index=True, right_index=True)

words = words.loc[~words.word.isna()]

list_morph = []
datalist = words.word.drop_duplicates()

# преобразование слов в начальную форму с наименованием частей речи
def add_morph(words_list):
    morph = pm.MorphAnalyzer()

    for word in log_progress(words_list, every = 1):
        try:
            m = morph.parse(word.strip())[0]
            list_morph.append({'normal_sentens':word, 'new':m.normal_form, 'tag':m.tag, 'grammema':m.tag.POS})
        except:
            continue

morph = pm.MorphAnalyzer()
# только нормальные граммемы
g = ['NOUN','ADJF', 'ADJS', 'COMP','VERB', 'INFN', 'PRTF', 'PRTS', 'GRND', 'ADVB']

bigs  = []
datalist = list(zip(df.normal_sentense, df.index_review))

def get_bigramms(sentences):
    for sentence, index in log_progress(sentences, 1):
        s = re.sub("[^\w]", " ", sentence)
        bigrm = nltk.bigrams(s.lower().split(" "))
        for l in list(bigrm):
            if l[0] != '' and l[1]!='':
                m = morph.parse(l[0])[0].tag.POS
                m2 = morph.parse(l[1])[0].tag.POS
                if str(m) in g and str(m2) in g:
                    bigs.extend([[index, l, l[0], l[1], m, m2]])

d_sent = []
datalist = list(zip(df.normal_sentense, df.index_review))

if __name__== "__main__":
    start = time.time()
    freeze_support()
    pool = Pool()
    
    ITERATION_COUNT = cpu_count()-1

    count_per_iteration = len(datalist) / float(ITERATION_COUNT)

    for i in range(0, ITERATION_COUNT):
        list_start = int(count_per_iteration * i)
        list_end = int(count_per_iteration * (i+1))
        pool.apply_async(get_sentiment, [datalist[list_start:list_end]])

b_df = pd.DataFrame(bigs).rename(columns = {0:'index_review', 1:'bigramm', 2:'word1', 3:'word2', 4:'tag1', 5:'tag2'})

b_df = bigramms.merge(df_temp[['index_review', 'url', 'score', 'positive', 'negative', 'comparative']], how = 'left', on = 'index_review').merge(places_df[['category', 'url']], how = 'left', on = 'url')

b_df

b_df.to_csv('../data/bigramms.csv', index = False)

len(datalist), len(list_morph)

words = words.merge(pd.DataFrame(list_morph).rename(columns = {'normal_sentens':'word', 'new':'normal_form'}), how = 'left', on = 'word')

words = words.merge(df[['index_review', 'url']], how = 'left', on = 'index_review')

words = words.merge(places_df[['category', 'url']], how = 'left', on = 'url')

words = words.merge(df_temp[['index_review', 'score', 'positive', 'negative', 'comparative']], how = 'left', on = 'index_review')

words.to_csv('../data/words.csv', index = False)

df['normal_sentense'] = df.index_review.apply(lambda x: ' '.join(words.loc[words.index_review==x, 'normal_form']))

# Определение сентиментов

df['text'] = df['text'].str.replace('ё', 'е')
df['normal_sentense'] = df['normal_sentense'].str.replace('ё', 'е')

df.loc[df.text.str.contains('ё')]

# код для семантического анализа

import re
import os
import csv
from collections import defaultdict


class Sentimental(object):
    def __init__(self, word_list=None, negation=None):
        self.word_list = {}
        self.negations = set()

        for wl_filename in self.__to_arg_list(word_list):
            self.load_word_list(wl_filename)
        for negations_filename in self.__to_arg_list(negation):
            self.load_neagations(negations_filename)

        self.__negation_skip = {'a', 'an', 'so', 'too', 'я', 'мне'}

    @staticmethod
    def __to_arg_list(obj):
        if obj is not None:
            if not isinstance(obj, list):
                obj = [obj]
        else:
            obj = []
        return obj

    def __is_prefixed_by_negation(self, token_idx, tokens):
        #   True if i != 0 and tokens[i - 1] in self.negations else False
        prev_idx = token_idx - 1
        if tokens[prev_idx] in self.__negation_skip:
            prev_idx -= 1

        is_prefixed = False
        if token_idx > 0 and prev_idx >= 0 and tokens[prev_idx] in self.negations:
            is_prefixed = True

        return is_prefixed

    def load_neagations(self, filename):
        with open(filename,'r', encoding='utf-8') as f:
            reader = csv.DictReader(f)
            negations = set([row['token'] for row in reader])
        self.negations |= negations

    def load_word_list(self, filename):
        with open(filename, 'r', encoding='utf-8') as f:
            reader = csv.DictReader(f)
            word_list = {row['word']: float(row['score']) for row in reader}
        self.word_list.update(word_list)

    def analyze(self, sentence):
        sentence_clean = re.sub(r'[^\w ]', ' ', sentence.lower())
        tokens = sentence_clean.split()

        scores = defaultdict(float)
        words = defaultdict(list)
        comparative = 0

        for i, token in enumerate(tokens):
            is_prefixed_by_negation = self.__is_prefixed_by_negation(i, tokens)
            if token in self.word_list and not is_prefixed_by_negation:
                score = self.word_list[token]

                score_type = 'negative' if score < 0 else 'positive'
                scores[score_type] += score
                words[score_type].append(token)

        if len(tokens) > 0:
            comparative = (scores['positive'] + scores['negative']) / len(tokens)

        result = {
            'score': scores['positive'] + scores['negative'],
            'positive': scores['positive'],
            'negative': scores['negative'],
            'comparative': comparative,
        }

        return result
sent = Sentimental(word_list=['D:/projects/word_list/afinn.csv', 'D:/projects/word_list/russian.csv'], negation='D:/projects/word_list/negations.csv')

def get_sentiment(sentences):
    for sentence, index in log_progress(sentences, 1):
        try:
            analyze = sent.analyze(sentence)
            list_sent = {}
            for attr in ['score', 'positive', 'negative', 'comparative']:
                list_sent.update({attr: analyze.get(attr)})
            list_sent.update({'index_review':index})
            d_sent.append(list_sent)
        except:
            print(index)

df_temp = df.merge(pd.DataFrame(d_sent), how = 'left', on = 'index_review').copy()

df_temp[['rating', 'positive', 'negative', 'score', 'text']].sample(6)

df.shape, pd.DataFrame(d_sent).shape

df_temp.loc[df_temp.score.isna()]

df_temp.loc[(df_temp.rating==1) &(df_temp.score>20)]

sent.analyze('отличный торговый центр гигантский ассортимент многий вид еда который нет в типичный российский магазин мочь сказать о это только положительный …')

get_sentiment(datalist[:2])

df_temp.to_csv('../data/comments_filled_sent.csv', index = False)
