==> notebooks/8 - Test the accuracy.ipynb <==
https://towardsdatascience.com/a-beginners-guide-to-sentiment-analysis-in-python-95e354ea84f6

import pandas as pd #работа с датафреймами- таблицами
import pymorphy2 as pm #части речи, формы
import re #регулярные выражения
from multiprocessing.dummy import Pool
from multiprocessing import cpu_count, freeze_support
import time
from IPython.display import display 
import nltk
import numpy as np

df = pd.read_csv('../data/comments_filled_sent.csv')

df = df.loc[~df.normal_sentense.isna()]

df.shape

df = df[df['rating'] != 3]
df['sentiment'] = df['rating'].apply(lambda rating : +1 if rating > 3 else -1)

# df = df[df['score'] > 0]
df['sentiment'] = df['score'].apply(lambda rating : 1 if rating > 0 else -1)

index = df.index
df['random_number'] = np.random.randn(len(index))
train = df[df['random_number'] <= 0.8]
test = df[df['random_number'] > 0.8]

from sklearn.feature_extraction.text import CountVectorizer
vectorizer = CountVectorizer(token_pattern=r'\b\w+\b')
train_matrix = vectorizer.fit_transform(train['normal_sentense'])
test_matrix = vectorizer.transform(test['normal_sentense'])

# Logistic Regression
from sklearn.linear_model import LogisticRegression
lr = LogisticRegression()

X_train = train_matrix
X_test = test_matrix
y_train = train['sentiment']
y_test = test['sentiment']

lr.fit(X_train,y_train)

predictions = lr.predict(X_test)

# find accuracy, precision, recall:
from sklearn.metrics import confusion_matrix,classification_report
new = np.asarray(y_test)
confusion_matrix(predictions,y_test)

print(classification_report(predictions,y_test))
